Main changes for Ibis 2 compatibility
