+++ ibis_bigquery/client.py
@@ -18,14 +18,10 @@ import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 import pandas as pd
-import regex as re
 from google.api_core.client_info import ClientInfo
-from google.api_core.exceptions import NotFound
-from ibis.client import Database, Query, SQLClient
+from ibis.client import Database, Query
 from multipledispatch import Dispatcher
-from pkg_resources import parse_version
 
-from . import compiler as comp
 from .datatypes import ibis_type_to_bigquery_type
 
 NATIVE_PARTITION_COL = "_PARTITIONTIME"
@@ -358,188 +354,3 @@ def parse_project_and_dataset(
         billing_project = data_project = project
 
     return data_project, billing_project, dataset
-
-
-class BigQueryClient(SQLClient):
-    """An ibis BigQuery client implementation."""
-
-    def __init__(
-        self,
-        backend,
-        project_id,
-        dataset_id=None,
-        credentials=None,
-        application_name=None,
-        partition_column=None,
-    ):
-        """Construct a BigQueryClient.
-
-        Parameters
-        ----------
-        project_id : str
-            A project name
-        dataset_id : Optional[str]
-            A ``<project_id>.<dataset_id>`` string or just a dataset name
-        credentials : google.auth.credentials.Credentials
-        application_name : str
-            A string identifying your application to Google API endpoints.
-        partition_column : str
-            Identifier to use instead of default ``_PARTITIONTIME`` partition
-            column.
-
-        """
-        self.query_class = backend.query_class
-        self.database_class = backend.database_class
-        self.table_class = backend.table_class
-        self.dialect = backend.dialect
-        (
-            self.data_project,
-            self.billing_project,
-            self.dataset,
-        ) = parse_project_and_dataset(project_id, dataset_id)
-        self.client = bq.Client(
-            project=self.billing_project,
-            credentials=credentials,
-            client_info=_create_client_info(application_name),
-        )
-        self.partition_column = partition_column
-
-    def _parse_project_and_dataset(self, dataset):
-        if not dataset and not self.dataset:
-            raise ValueError("Unable to determine BigQuery dataset.")
-        project, _, dataset = parse_project_and_dataset(
-            self.billing_project,
-            dataset or "{}.{}".format(self.data_project, self.dataset),
-        )
-        return project, dataset
-
-    @property
-    def project_id(self):
-        return self.data_project
-
-    @property
-    def dataset_id(self):
-        return self.dataset
-
-    def table(self, name, database=None):
-        t = super().table(name, database=database)
-        project, dataset, name = t.op().name.split(".")
-        dataset_ref = self.client.dataset(dataset, project=project)
-        table_ref = dataset_ref.table(name)
-        bq_table = self.client.get_table(table_ref)
-        return rename_partitioned_column(t, bq_table, self.partition_column)
-
-    def _build_ast(self, expr, context):
-        result = comp.build_ast(expr, context)
-        return result
-
-    def _get_query(self, dml, **kwargs):
-        return self.query_class(self, dml, query_parameters=dml.context.params)
-
-    def _fully_qualified_name(self, name, database):
-        project, dataset = self._parse_project_and_dataset(database)
-        return "{}.{}.{}".format(project, dataset, name)
-
-    def _get_table_schema(self, qualified_name):
-        dataset, table = qualified_name.rsplit(".", 1)
-        assert dataset is not None, "dataset is None"
-        return self.get_schema(table, database=dataset)
-
-    def _get_schema_using_query(self, limited_query):
-        with self._execute(limited_query, results=True) as cur:
-            # resets the state of the cursor and closes operation
-            names, ibis_types = self._adapt_types(cur.description)
-        return sch.Schema(names, ibis_types)
-
-    def _adapt_types(self, descr):
-        names = []
-        adapted_types = []
-        for col in descr:
-            names.append(col.name)
-            typename = bigquery_field_to_ibis_dtype(col)
-            adapted_types.append(typename)
-        return names, adapted_types
-
-    def _execute(self, stmt, results=True, query_parameters=None):
-        job_config = bq.job.QueryJobConfig()
-        job_config.query_parameters = query_parameters or []
-        job_config.use_legacy_sql = False  # False by default in >=0.28
-        query = self.client.query(
-            stmt, job_config=job_config, project=self.billing_project
-        )
-        query.result()  # blocks until finished
-        return BigQueryCursor(query)
-
-    def database(self, name=None):
-        if name is None and self.dataset is None:
-            raise ValueError(
-                "Unable to determine BigQuery dataset. Call "
-                "client.database('my_dataset') or set_database('my_dataset') "
-                "to assign your client a dataset."
-            )
-        return self.database_class(name or self.dataset, self)
-
-    @property
-    def current_database(self):
-        return self.database(self.dataset)
-
-    def set_database(self, name):
-        self.data_project, self.dataset = self._parse_project_and_dataset(name)
-
-    def exists_database(self, name):
-        project, dataset = self._parse_project_and_dataset(name)
-        client = self.client
-        dataset_ref = client.dataset(dataset, project=project)
-        try:
-            client.get_dataset(dataset_ref)
-        except NotFound:
-            return False
-        else:
-            return True
-
-    def list_databases(self, like=None):
-        results = [
-            dataset.dataset_id
-            for dataset in self.client.list_datasets(project=self.data_project)
-        ]
-        if like:
-            results = [
-                dataset_name
-                for dataset_name in results
-                if re.match(like, dataset_name) is not None
-            ]
-        return results
-
-    def exists_table(self, name, database=None):
-        project, dataset = self._parse_project_and_dataset(database)
-        client = self.client
-        dataset_ref = self.client.dataset(dataset, project=project)
-        table_ref = dataset_ref.table(name)
-        try:
-            client.get_table(table_ref)
-        except NotFound:
-            return False
-        else:
-            return True
-
-    def list_tables(self, like=None, database=None):
-        project, dataset = self._parse_project_and_dataset(database)
-        dataset_ref = bq.DatasetReference(project, dataset)
-        result = [table.table_id for table in self.client.list_tables(dataset_ref)]
-        if like:
-            result = [
-                table_name
-                for table_name in result
-                if re.match(like, table_name) is not None
-            ]
-        return result
-
-    def get_schema(self, name, database=None):
-        project, dataset = self._parse_project_and_dataset(database)
-        table_ref = self.client.dataset(dataset, project=project).table(name)
-        bq_table = self.client.get_table(table_ref)
-        return sch.infer(bq_table)
-
-    @property
-    def version(self):
-        return parse_version(bq.__version__)
